+++ internal/provider/vpcs_data_source.go
@@ -24,21 +24,6 @@ func NewVpcsDataSource() datasource.DataSource {
 	return &vpcsDataSource{}
 }
 
-var (
-	PeeringConnectionsDSType = types.ObjectType{
-		AttrTypes: map[string]attr.Type{
-			"vpc_id":           types.StringType,
-			"provisioned_id":   types.StringType,
-			"status":           types.StringType,
-			"error_message":    types.StringType,
-			"peer_vpc_id":      types.StringType,
-			"peer_cidr":        types.StringType,
-			"peer_account_id":  types.StringType,
-			"peer_region_code": types.StringType,
-		},
-	}
-)
-
 // vpcsDataSource is the data source implementation.
 type vpcsDataSource struct {
 	client *tsClient.Client
@@ -72,6 +57,7 @@ type peeringConnectionDSModel struct {
 	Status         types.String `tfsdk:"status"`
 	ErrorMessage   types.String `tfsdk:"error_message"`
 	PeerVPCID      types.String `tfsdk:"peer_vpc_id"`
+	PeerCIDRBlocks types.List   `tfsdk:"peer_cidr_blocks"`
 	PeerCIDR       types.String `tfsdk:"peer_cidr"`
 	PeerAccountID  types.String `tfsdk:"peer_account_id"`
 	PeerRegionCode types.String `tfsdk:"peer_region_code"`
@@ -125,6 +111,14 @@ func (d *vpcsDataSource) Read(ctx context.Context, _ datasource.ReadRequest, res
 			pcm.Status = types.StringValue(pc.Status)
 			pcm.PeerVPCID = types.StringValue(pc.PeerVPC.ID)
 			pcm.PeerAccountID = types.StringValue(pc.PeerVPC.AccountID)
+
+			peerCIDRBlocks, cidrDiags := types.ListValueFrom(ctx, types.StringType, pc.PeerVPC.CIDRBlocks)
+			resp.Diagnostics.Append(cidrDiags...)
+			if resp.Diagnostics.HasError() {
+				return
+			}
+			pcm.PeerCIDRBlocks = peerCIDRBlocks
+
 			pcm.PeerCIDR = types.StringValue(pc.PeerVPC.CIDR)
 			pcm.PeerRegionCode = types.StringValue(pc.PeerVPC.RegionCode)
 			pcms = append(pcms, pcm)
@@ -201,6 +195,7 @@ func (d *vpcsDataSource) Schema(_ context.Context, _ datasource.SchemaRequest, r
 									"status":           types.StringType,
 									"error_message":    types.StringType,
 									"peer_vpc_id":      types.StringType,
+									"peer_cidr_blocks": types.ListType{ElemType: types.StringType},
 									"peer_cidr":        types.StringType,
 									"peer_account_id":  types.StringType,
 									"peer_region_code": types.StringType,